requirements.txt
discord.py[voice]
async-timeout
requests
beautifulsoup4
yt-dlp==2021.3.7
discord.py==2.2.2
ffmpeg
youtube-search
asyncpraw
pytube == 12.1.3
spotify_dl
